refactoring: by review and add minimum test
