+++ mainLogic/src/main/java/ru/practicum/events/controller/user/EventController.java
@@ -8,11 +8,12 @@ import ru.practicum.events.dto.LocationShort;
 import ru.practicum.events.model.Address;
 import ru.practicum.events.model.UpdateEventRequest;
 import ru.practicum.events.service.user.EventService;
-import ru.practicum.events.service.user.LocationService;
+import ru.practicum.geocoding.geoService.LocationService;
 
 
 import javax.servlet.http.HttpServletRequest;
 import java.util.List;
+import java.util.Optional;
 
 @Slf4j
 @RestController
@@ -21,12 +22,10 @@ public class EventController {
 
     private final EventService eventService;
 
-    private final LocationService locationService;
 
 
-    public EventController(EventService eventService, LocationService locationService) {
+    public EventController(EventService eventService) {
         this.eventService = eventService;
-        this.locationService = locationService;
     }
 
     @PostMapping("/{userId}/events")
@@ -70,42 +69,22 @@ public class EventController {
         return eventService.cansel(userId, eventId);
     }
 
-    @GetMapping("/nearby/{lon}/{lat}/{distance}")
-    public List<EventDto> findNearby(@PathVariable(name = "lat") float lat,
-                                     @PathVariable(name = "lon") float lon,
-                                     @PathVariable(name = "distance") float distance,
-                                     HttpServletRequest request) {
-        log.info(request.getRequestURI() + " " + request.getQueryString() + " " + request.getMethod());
-        return eventService.searchFindNearbyByCoordinate(lat, lon, distance);
-    }
 
     @GetMapping("/nearby")
-    public List<EventDto> findNearbyByAddress(@RequestParam(name = "city") String city,
-                                              @RequestParam(name = "street") String street,
-                                              @RequestParam(name = "number") String number,
-                                              @RequestParam(name = "distance") Float distance,
+    public List<EventDto> findNearbyByAddress(@RequestParam(name = "city") Optional<String > city,
+                                              @RequestParam(name = "street") Optional<String> street,
+                                              @RequestParam(name = "number") Optional<String> number,
+                                              @RequestParam(name = "lat") Optional<Float> lat,
+                                              @RequestParam(name = "lon") Optional<Float> lon,
+                                              @RequestParam(name = "distance", defaultValue = "1") float distance,
                                               HttpServletRequest request) {
         log.info(request.getRequestURI() + " " + request.getQueryString() + " " + request.getMethod());
-        return eventService.findNearbyByAddress(city, street, number, distance);
+
+        return ((city.isPresent()&&street.isPresent()&&number.isPresent()) ?
+                eventService.findNearbyByAddress(city.get(), street.get(), number.get(), distance) :
+                eventService.searchFindNearbyByCoordinate(lat, lon, distance));
     }
 
-
-    @GetMapping("/getlocation")
-    public LocationShort getLocationByAddress(@RequestParam(name = "city") String city,
-                                              @RequestParam(name = "street") String street,
-                                              @RequestParam(name = "number") String number,
-                                              HttpServletRequest request) {
-        log.info(request.getRequestURI() + " " + request.getQueryString() + " " + request.getMethod());
-        return locationService.toCoordinate(city, street, number);
-    }
-
-    @GetMapping("/getaddress")
-    public Address getAddress(@RequestParam(name = "lon") float lon,
-                              @RequestParam(name = "lat") float lat,
-                              HttpServletRequest request) {
-        log.info(request.getRequestURI() + " " + request.getQueryString() + " " + request.getMethod());
-        return locationService.toAddress(lon, lat);
-    }
 
 
 }
